+++ app/src/js/components/Executions/overview.js
@@ -11,7 +11,6 @@ import {
   listExecutions,
   listWorkflows,
   getOptionsCollectionName,
-  refreshCumulusDbConnection,
 } from '../../actions';
 import { workflowOptions as workflowSelectOptions } from '../../selectors';
 import { tally, lastUpdated } from '../../utils/format';
@@ -50,7 +49,6 @@ const ExecutionOverview = ({
   const tableColumnsArray = tableColumns({ dispatch });
 
   useEffect(() => {
-    dispatch(refreshCumulusDbConnection());
     dispatch(listWorkflows());
   }, [dispatch]);