+++ app/src/js/components/ReconciliationReports/list.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Link, withRouter } from 'react-router-dom';
@@ -8,7 +8,6 @@ import {
   clearReconciliationReportsFilter,
   listReconciliationReports,
   filterReconciliationReports,
-  refreshCumulusDbConnection,
 } from '../../actions';
 import { lastUpdated } from '../../utils/format';
 import { reconciliationReportStatus as statusOptions } from '../../utils/status';
@@ -63,10 +62,6 @@ const ReconciliationReportList = ({
   const { queriedAt, count } = list.meta;
   const query = generateQuery();
   const tableColumnsArray = tableColumns({ dispatch, isGranules, query });
-
-  useEffect(() => {
-    dispatch(refreshCumulusDbConnection());
-  }, [dispatch]);
 
   function generateQuery() {
     return {